feat(wizard) moved dashboard components back into general components folder
AB#57
