+++ Singer.API/ClientApp/package.json
@@ -1,6 +1,6 @@
 {
    "name": "Singer",
-   "version": "0.1.0-issue-AB-57-family-registration-wizard.882",
+   "version": "0.1.0-issue-AB-57-family-registration-wizard.887",
    "scripts": {
       "ng": "ng",
       "start": "ng serve",
