+++ Singer.API/ClientApp/src/app/modules/core/services/singerevents-api/singerevents.service.ts
@@ -16,7 +16,7 @@ import {
    UserRegisteredDTO,
 } from '../../models/event-registration.model';
 import { TimeUnit, RepeatType } from '../../models/enum';
-import { SearchEventData } from 'src/app/modules/dashboard/components/shared-components/event-search/event-search.component';
+import { SearchEventData } from 'src/app/modules/dashboard/components/event-search/event-search.component';
 
 @Injectable({
    providedIn: 'root',
