+++ Singer.API/ClientApp/src/app/modules/dashboard/components/about/about.component.html
