+++ Singer.API/ClientApp/src/app/modules/dashboard/components/admin-menu/admin-menu.component.css
