+++ Singer.API/ClientApp/src/app/modules/dashboard/components/event-card/event-card.component.ts
