+++ Singer.API/ClientApp/src/app/modules/dashboard/components/event-list/event-list.component.css
