+++ Singer.API/ClientApp/src/app/modules/dashboard/components/event-search/event-search.component.ts
