+++ Singer.API/ClientApp/src/app/modules/dashboard/components/home/home.component.html
