+++ Singer.API/ClientApp/src/app/modules/dashboard/dashboard.module.ts
@@ -3,17 +3,17 @@ import { CommonModule } from '@angular/common';
 import { DashboardComponent } from './pages/dashboard/dashboard.component';
 import { MaterialModule } from 'src/app/material.module';
 import { DashboardRoutingModule } from './dashboard-routing.module';
-import { HomeComponent } from './components/shared-components/home/home.component';
-import { AboutComponent } from './components/shared-components/about/about.component';
+import { HomeComponent } from './components/home/home.component';
+import { AboutComponent } from './components/about/about.component';
 import { RouterModule } from '@angular/router';
-import { EventCardComponent } from './components/shared-components/event-card/event-card.component';
-import { EventListComponent } from './components/shared-components/event-list/event-list.component';
+import { EventCardComponent } from './components/event-card/event-card.component';
+import { EventListComponent } from './components/event-list/event-list.component';
 import { CoreModule } from '../core/core.module';
-import { EventSearchComponent } from './components/shared-components/event-search/event-search.component';
+import { EventSearchComponent } from './components/event-search/event-search.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ApiService } from '../core/services/api.service';
 import { SharedModule } from '../shared/shared.module';
-import { AdminMenuComponent } from './components/admin-components/admin-menu/admin-menu.component';
+import { AdminMenuComponent } from './components/admin-menu/admin-menu.component';
 
 @NgModule({
    declarations: [
